data_pipeline/requirements.txt
psycopg2-binary>=2.9
requests>=2.28
FinMind>=0.7
